o merge in changes from trunk
git-svn-id: https://svn.apache.org/repos/asf/maven/artifact/branches/CAP@659009 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/maven/artifact/repository/ArtifactRepositoryPolicy.java
@@ -131,7 +131,7 @@ public class ArtifactRepositoryPolicy
         else if ( updatePolicy.startsWith( UPDATE_POLICY_INTERVAL ) )
         {
             String s = updatePolicy.substring( UPDATE_POLICY_INTERVAL.length() + 1 );
-            int minutes = Integer.valueOf( s ).intValue();
+            int minutes = Integer.valueOf(s);
             Calendar cal = Calendar.getInstance();
             cal.add( Calendar.MINUTE, -minutes );
             if ( cal.getTime().after( lastModified ) )
